fix: resolve Psalm strict operand type errors
Cast divisors to float to satisfy Psalm strict mode:
- ErrorSuggestionHelper: use 100.0 instead of 100
- Profiler: cast count() results to float for division

Note: 59 Psalm errors remain in new code (mostly type coercion warnings).
These can be addressed in follow-up commits or by adjusting Psalm baseline.

+++ src/Framework/Profiler/Profiler.php
@@ -136,7 +136,7 @@ final class Profiler
 
         foreach ($byOperation as $operation => $stats) {
             $byOperation[$operation]['avg_duration'] = $stats['count'] > 0
-                ? round($stats['total_duration'] / $stats['count'], 6)
+                ? round($stats['total_duration'] / (float) $stats['count'], 6)
                 : 0.0;
         }
 
@@ -144,7 +144,7 @@ final class Profiler
             'total_operations' => count($this->entries),
             'total_duration' => round($totalDuration, 6),
             'avg_duration' => count($this->entries) > 0
-                ? round($totalDuration / count($this->entries), 6)
+                ? round($totalDuration / (float) count($this->entries), 6)
                 : 0.0,
             'peak_memory' => $peakMemory,
             'by_operation' => $byOperation,
